backend/requirements.txt
Flask
flask-cors
PyJWT
pymongo
dnspython